moved file selection in cogShop file class
